extract the catefgory interface
