+++ src/main/java/spring/workshop/expenses/services/CategoryServiceImpl.java
@@ -1,0 +1,71 @@
+package spring.workshop.expenses.services;
+
+import java.util.Comparator;
+import java.util.List;
+import java.util.Optional;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.stereotype.Service;
+
+import spring.workshop.expenses.entities.Category;
+import spring.workshop.expenses.repos.CategoryRepository;
+
+@Service
+public class CategoryServiceImpl implements CategoryService {
+    private CategoryRepository categoryRepository;
+    private static final Logger LOG = LoggerFactory.getLogger(CategoryServiceImpl.class);
+
+    public CategoryServiceImpl(CategoryRepository categoryRepository) {
+        this.categoryRepository = categoryRepository;
+    }
+
+    @Override
+    public void updateCategory(Integer id, Category category) {
+        Optional<Category> oldCategory = categoryRepository.findById(id);
+        if (oldCategory.isPresent()) {
+            Category updatedCategory = oldCategory.get();
+            updatedCategory.setName(category.getName());
+            categoryRepository.save(updatedCategory);
+            LOG.info("Category {} updated", category.getName());
+        } else {
+            LOG.warn("Category {} not found for update", category.getName());
+            throw new IllegalArgumentException("No such category with id " + id);
+        }
+    }
+
+    @Override
+    public void deleteCategory(Integer id) {
+        Optional<Category> category = categoryRepository.findById(id);
+        if (category.isPresent()) {
+            categoryRepository.delete(category.get());
+            LOG.info("Category {} deleted", category.get().getName());
+        } else {
+            LOG.warn("Category ID {} not found for delete", id);
+            throw new IllegalArgumentException("No such category with id " + id);
+        }
+    }
+
+    @Override
+    public Category addCategory(Category category) {
+        Category newCategory = categoryRepository.save(category);
+        LOG.info("Category {} added", category.getName());
+        return newCategory;
+    }
+
+    @Override
+    public List<Category> findAll() {
+        List<Category> categories = categoryRepository.findAll();
+        categories.sort(Comparator.comparing(Category::getId));
+        return categories;
+    }
+
+    @Override
+    public Category findById(Integer id) {
+        Optional<Category> category = categoryRepository.findById(id);
+        if (!category.isPresent())
+            throw new IllegalArgumentException("No such category with id " + id);
+
+        return category.get();
+    }
+}
